Listar las valoraciones de recursos y los recursos de las listas de recursos.

+++ src/main/java/com/grupo1/recursos_tic/controller/ResourceController.java
@@ -1,6 +1,8 @@
 package com.grupo1.recursos_tic.controller;
 
 import com.grupo1.recursos_tic.model.Resource;
+import com.grupo1.recursos_tic.model.ResourceList;
+import com.grupo1.recursos_tic.service.RatingService;
 import com.grupo1.recursos_tic.service.ResourceService;
 
 import lombok.AllArgsConstructor;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import static com.grupo1.recursos_tic.util.Utility.*;
 
@@ -21,6 +24,7 @@ import static com.grupo1.recursos_tic.util.Utility.*;
 public class ResourceController {
 
     private ResourceService resourceService;
+    private RatingService ratingService;
 
     private final String idMsg = "Falta el id o no es un entero positivo";
     private final String notIdMsg = "El recurso no existe";
@@ -39,8 +43,12 @@ public class ResourceController {
         if (invalidIntPosNumber(id) || id == 0)
             throw new NoSuchElementException(idMsg);
 
+        if (!resourceService.existsById(id))
+            throw new NoSuchElementException(notIdMsg);
+
         return resourceService.findById(id).map(resource -> {
             model.addAttribute("resource", resource);
+            model.addAttribute("ratings", ratingService.findAllByResource_Id(id));
             return "resource/detail";
         }).orElseThrow(() -> new NoSuchElementException(notIdMsg));
     }
@@ -55,6 +63,9 @@ public class ResourceController {
     public String getFormToUpdate(Model model, @PathVariable Long id) {
         if (invalidIntPosNumber(id) || id == 0)
             throw new NoSuchElementException(idMsg);
+
+        if (!resourceService.existsById(id))
+            throw new NoSuchElementException(notIdMsg);
 
         return resourceService.findById(id).map(resource -> {
             model.addAttribute("resource", resource);
@@ -84,12 +95,18 @@ public class ResourceController {
     public String deleteById(Model model, @PathVariable Long id) {
         if (invalidIntPosNumber(id) || id == 0)
             throw new NoSuchElementException(idMsg);
+
+        if (!resourceService.existsById(id))
+            throw new NoSuchElementException(notIdMsg);
+
         return resourceService.findById(id).map(resource -> {
             resourceService.deleteById(resource.getId());
             return "redirect:/resources";
         }).orElseThrow(() -> new NoSuchElementException(notIdMsg));
     }
 
+    // TODO Sólo el admin puede borrar recursos
+    // TODO Eliminar recursos no funciona
     @GetMapping("resources/delete")
     public String deleteAll(Model model) {
         resourceService.deleteAll();
